enterdb_ldb_worker:recreate_shard/1 makes cast.
Previously recreate_shard was making a gen_server:call call but timeout may occur which cause crash on wrapping tables.
Now recreate_shard is making a gen_serveir:cast call and actual recreation process can run on background. (Shard will not be available for database operations during recreation but it shouldn't be an issue since these shards are the least likely to be used).